Security: sqlite-ndp
This commit introduces support for a version of sqlite-ndp that
can read encrypted databases. This is based on sqlcipher. This
ensures that the confidentiality and integrity of the data in the
sqlite database is maintained/ensured at the granularity of a page.

This commit also solves a synchronization bug that prevented the
consumer thread from knowing whether the producer thread has completed
its job.

Signed-off-by: Harshavardhan Unnibhavi <harshanavkis@gmail.com>